compiler-rt: crt0: all memory polyfills accept and return u128 (#196)
This approach, simialr to how ALU polyfills are defined, makes it easier
to write the Rust part of Hieratika.

+++ compiler-rt/src/crt0/memory/alloca.cairo
@@ -4,6 +4,6 @@ use super::alloc::__llvm_alloc_l_l_p;
 
 /// Due to how Cairo memory model works, there is no difference between a stack allocation and a
 /// heap allocation. Therefore, polyfills for `alloca` are just wrappers over `alloc`.
-pub fn __llvm_alloca_l_l_p(ref state: RTState, size_bytes: i64, count: i64) -> Address {
+pub fn __llvm_alloca_l_l_p(ref state: RTState, size_bytes: u128, count: u128) -> Address {
     __llvm_alloc_l_l_p(ref state, size_bytes, count)
 }
